feat: validate numbers formatted for the current locale
NumberType, IntegerType, MoneyType and PercentType render their value
through the ICU number formatter, so an "it" form holds "12,5" where the
model holds 12.5. There were no JavaScript counterparts for their view
transformers, and an unknown transformer is skipped silently, so the raw
localized string reached the constraints and every numeric check failed.

The factory now exports the formatting rules of those transformers: the
decimal and grouping separators, the grouping flag, the scale and the
rounding mode. They are read from the formatter the transformer itself
builds, so options that change the formatting are honoured, including an
IntegerType without grouping, which Symfony renders with the "en" locale
regardless of the current one.

Four transformers reverse those rules in JavaScript on a shared parser
that follows the ICU behaviour: both separators are accepted where that
is not ambiguous, groups must hold three digits, and all seven rounding
modes are applied on the scale of the field.

A value that cannot be reversed is now reported with the invalid_message
of the element and skips its constraints, the way Symfony reports a
failed transformation. Until now an exception from a transformer simply
broke the validation of the form.

Range also parsed its bounds with parseInt, which truncated decimal
limits such as min: 1.5 to 1.

Closes formapro/JsFormValidatorBundle#140

+++ src/Resources/public/js/constraints/Range.js
@@ -54,8 +54,8 @@ export default function SymfonyComponentValidatorConstraintsRange() {
     };
 
     this.onCreate = function () {
-        this.min = parseInt(this.min);
-        this.max = parseInt(this.max);
+        this.min = parseFloat(this.min);
+        this.max = parseFloat(this.max);
     }
 }
